+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Deprecated
 ### Removed
 ### Fixed
+- [ViGEm Provider] Fix Issue #41  
+No longer crashes if ViGEm not installed
 
 ## 0.11.1 - 2019-12-15
 ### Added